Updated CardData and enums
There is now one CardType enum, that holds values for both SurfaceCards and OutpostCards.
Keywords are now "Traits", and it has been extended so that it includes both those for SurfaceCards and for OutpostCards. The types and variable names for this have been updated in all places now, including methodnames and such.

There is a new enum called keywords however, but that has other uses.

I also cleaned up the Ideology enum slightly to make it easier to read.