Added delegate to get cardObjects based on filter
Added a GetCardMatching() function that takes a delegate function and returns a list of cards that matches. This way I should be able to use lambda expression to get lists based on keyword, cardtype, trait as well as any combinations of these.
Also made the list of cards in CardZone private to ensure that it will handle its own adding and removal from now on.