Working on getting collected values from zones
Will continue tomorrow to see how I best get the collected values from a zone, and how I then ask the slot for the appropriate card. Will also need to check if I have run in to a potential error when it comes to the cardscript not inheriting from MonoBehavior.